# design_&_architecture/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: WebApp Value Investing API
  description: |
    REST API per la WebApp Value Investing. Espone analisi quantitative Graham/Buffett
    su titoli quotati USA + watchlist personale.
    Contratto consumato dal frontend Next.js (SPA). Vedi overview architetturale in
    `design_&_architecture/overview.md`.
  version: "1.0.0-MVP"
  contact:
    name: Lead Architect
servers:
  - url: http://localhost:8080
    description: dev locale
  - url: https://api.example.com
    description: prod (placeholder)
...
paths:
...
  /api/analysis/{ticker}/deep:
    get:
      tags: [analysis]
      summary: Deep analysis completa — ROE + rule engine + verdict + Munger + sentiment (US-045, ADR-020)
      description: |
        Orchestrazione deep analysis pipeline: filing ingest → embedding → ROE
        dual-lookback → rule engine 13 ruleId → price action → verdict cascade →
        position sizing. I campi LLM-dependent (mungerReport, newsSentiment) sono
        popolati solo se invoke_llm=true; in default restano null con
        llmStatus="NOT_INVOKED".
      parameters:
        - $ref: '#/components/parameters/Ticker'
        - name: invoke_llm
          in: query
          required: false
          description: |
            Enable LLM-dependent blocks (Munger inversion report, news sentiment).
            Default false — deterministic-only response in < 2s.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Deep analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepAnalysisResponse'
        '404':
          description: Ticker non trovato
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '422':
          description: Ticker senza filing SEC disponibili
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ProblemDetails'
                  - type: object
                    properties:
                      reason:
                        type: string
                        enum: [no_sec_filings]
        '503':
          description: FMP o LLM indisponibile
          content:
            application/problem+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ProblemDetails'
                  - type: object
                    properties:
                      reason:
                        type: string
                        enum: [llm_unavailable, fmp_unavailable]
  /api/analysis/{ticker}/deep/runs:
    post:
      tags: [analysis]
      summary: Avvia un'esecuzione ASINCRONA della deep analysis (US-045)
      description: |
        Avvio fire-and-enqueue della deep analysis pipeline. Risponde 202 Accepted
        immediatamente senza attendere il completamento; l'esecuzione prosegue in
        background (executor @Async) e popola la tabella deep_analysis_run.
        Dedupe: se per il ticker esiste già un run in stato RUNNING, NON ne avvia
        uno nuovo e restituisce il runId di quello in corso. L'esito finale si
        recupera via GET /api/analysis/{ticker}/deep/latest. L'endpoint sincrono
        GET /api/analysis/{ticker}/deep resta disponibile per back-compat.
      parameters:
        - $ref: '#/components/parameters/Ticker'
        - name: invoke_llm
          in: query
          required: false
          description: |
            Enable LLM-dependent blocks (Munger inversion report, news sentiment)
            for the background run. Default false.
          schema: { type: boolean, default: false }
      responses:
        '202':
          description: Run avviato (o run RUNNING esistente riusato per dedupe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepAnalysisRunStatusResponse'
        '404':
          description: Ticker non trovato
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/analysis/{ticker}/deep/latest:
    get:
      tags: [analysis]
      summary: Ultima esecuzione (asincrona) della deep analysis per il ticker (US-045)
      description: |
        Restituisce l'ultima esecuzione persistita in deep_analysis_run per il
        ticker. All'apertura pagina si mostra questo stato senza ri-eseguire la
        pipeline (no auto-rerun). status=NONE quando non esiste ancora alcun run;
        result è popolato (DeepAnalysisResponse) solo quando status=SUCCESS;
        error è popolato solo quando status=FAILED.
      parameters:
        - $ref: '#/components/parameters/Ticker'
      responses:
        '200':
          description: Stato dell'ultima esecuzione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LatestDeepAnalysisResponse'
        '404':
          description: Ticker non trovato
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/analysis/{ticker}/deep/ingest:
    post:
      tags: [analysis]
      summary: Avvia l'INGEST ASINCRONO dei filing SEC (download + embedding) (US-045)
      description: |
        Avvio fire-and-enqueue dell'indicizzazione: scarica i filing SEC e calcola/
        salva gli embedding, SENZA invocare l'LLM. Risponde 202 Accepted
        immediatamente; l'esecuzione prosegue in background (stesso executor @Async,
        dispatch per kind) e popola la tabella deep_analysis_run con kind=INGEST.
        Idempotente: i filing già presenti in filing_chunks vengono saltati
        (chunksSkipped). Dedupe: se per il ticker esiste già un INGEST in stato
        RUNNING, NON ne avvia uno nuovo e restituisce il runId di quello in corso.
        Gli embedding persistiti sono poi riusati dall'analisi LLM
        (POST .../deep/runs?invoke_llm=true).
      parameters:
        - $ref: '#/components/parameters/Ticker'
      responses:
        '202':
          description: INGEST avviato (o INGEST RUNNING esistente riusato per dedupe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeepAnalysisRunStatusResponse'
        '404':
          description: Ticker non trovato
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /api/analysis/{ticker}/deep/ingest/latest:
    get:
      tags: [analysis]
      summary: Ultima indicizzazione (INGEST asincrono) dei filing per il ticker (US-045)
      description: |
        Restituisce l'ultimo INGEST persistito in deep_analysis_run (kind=INGEST)
        per il ticker. status=NONE quando non esiste ancora alcun INGEST; summary
        (IngestSummary) è popolato solo quando status=SUCCESS; error solo quando
        status=FAILED.
      parameters:
        - $ref: '#/components/parameters/Ticker'
      responses:
        '200':
          description: Stato dell'ultima indicizzazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestStatusResponse'
        '404':
          description: Ticker non trovato
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }

  # ---------------- EP-005 historical + moat ----------------
...
components:
  parameters:
    Ticker:
      name: ticker
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 10
        pattern: '^[A-Za-z0-9.\-]{1,10}$'
      description: Ticker simbolo (case-insensitive, normalizzato uppercase)
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details; extension members al top-level (ADR-012).
      properties:
        type:    { type: string, format: uri }
        title:   { type: string }
        status:  { type: integer }
        detail:  { type: string }
        instance: { type: string }
        ticker:    { type: string, description: "Extension — ticker correlato" }
        timestamp: { type: string, format: date-time, description: "Extension — istante risposta" }
        requestId: { type: string, description: "Extension — correlazione X-Request-Id" }
        reason:    { type: string, description: "Extension — codice motivo dominio" }
    Signal:
      type: string
      enum: [GREEN, YELLOW, RED, INDETERMINATE, NOT_APPLICABLE, NOT_CALCULABLE]
...
    RuleSignal:
      type: object
      required: [ruleId, signal, threshold]
      description: |
        Esito atomico di una singola regola del Rule Engine per un ticker.
        Il dettaglio strutturato per-rule (es. yearsPositive, lossYears, pe3yAvg, revenueLatest, ecc.)
        e' veicolato nel campo `rationale` (stringa human-readable) e in `observedValue` (numerico).
        La strutturazione tipata per-ruleId (oneOf discriminato / additionalProperties) e' rimandata
        a un futuro refactor del contratto: non scope di TSK-087/EP-010 per evitare breaking change
        sui 7 ruleId Buffett gia' in produzione e sui client TS auto-generati.
      properties:
        ruleId:
          type: string
          description: |
            Identificatore stabile della regola di valutazione. Set chiuso versionato a contratto.
            L'x-extension `x-buffett-quality` raggruppa le 7 regole Buffett-leaning (EP-003).
            L'x-extension `x-graham-defensive` raggruppa le 6 regole Graham defensive (EP-010,
            Cap.14 "The Defensive Investor and Common Stocks", Intelligent Investor).
          enum:
            - ROE_10Y_AVG
            - ROIC_10Y_AVG
            - GROSS_MARGIN_10Y_AVG
            - NET_MARGIN_10Y_AVG
            - CURRENT_RATIO_LATEST
            - DEBT_TO_INCOME_LATEST
            - CAPEX_INTENSITY_10Y_AVG
            - SIZE_LATEST
            - EARNINGS_STABILITY_10Y
            - EPS_GROWTH_10Y
            - PE_3Y_AVG
            - PB_LATEST
            - DIVIDEND_CONTINUITY_20Y
          example: ROE_10Y_AVG
          x-buffett-quality:
            - ROE_10Y_AVG
            - ROIC_10Y_AVG
            - GROSS_MARGIN_10Y_AVG
            - NET_MARGIN_10Y_AVG
            - CURRENT_RATIO_LATEST
            - DEBT_TO_INCOME_LATEST
            - CAPEX_INTENSITY_10Y_AVG
          x-graham-defensive:
            - SIZE_LATEST
            - EARNINGS_STABILITY_10Y
            - EPS_GROWTH_10Y
            - PE_3Y_AVG
            - PB_LATEST
            - DIVIDEND_CONTINUITY_20Y
        signal:        { $ref: '#/components/schemas/Signal' }
        observedValue: { type: number, nullable: true }
        threshold:     { type: string }
        rationale:     { type: string }
...
    VerdictClass:
      type: string
      enum:
        - APPROVATO_PANIC_BUY
        - APPROVATO
        - WATCHLIST
        - BOCCIATO_NUMERICO
        - BOCCIATO_QUALITATIVO
        - BOCCIATO_VALUE_TRAP
    LivelloRischio:
      type: string
      enum:
        - RISCHIO_BASSO
        - RISCHIO_MODERATO
        - RISCHIO_ALTO
        - RISCHIO_ESTREMO
    SentimentClass:
      type: string
      enum:
        - TEMPORARY_PANIC
        - STRUCTURAL_DAMAGE
        - NEUTRAL
    LlmStatus:
      type: string
      enum: [INVOKED, NOT_INVOKED, CACHE_HIT]
    RoeBlock:
      type: object
      required: [fiveYearDataPoints, tenYearDataPoints]
      description: |
        Dual-lookback ROE metrics (ADR-020). Both 5-year (growth/turnaround
        signal) and 10-year (Graham defensive stability signal) averages
        exposed side-by-side. Nullable avg fields handle IPO-recent tickers.
      properties:
        fiveYearAvg:
          type: number
          nullable: true
          description: 5-year average ROE as fraction (e.g. 0.265 = 26.5%)
        tenYearAvg:
          type: number
          nullable: true
          description: 10-year average ROE as fraction (e.g. 0.312 = 31.2%)
        fiveYearDataPoints:
          type: integer
          minimum: 0
          maximum: 5
        tenYearDataPoints:
          type: integer
          minimum: 0
          maximum: 10
    PriceActionBlock:
      type: object
      required: [panicDiscount, deteriorationWarning, seriesDays]
      description: Price action snapshot — 52-week range, moving averages, signals.
      properties:
        priceNow:              { type: number, nullable: true }
        max52w:                { type: number, nullable: true }
        min52w:                { type: number, nullable: true }
        drawdownPct:           { type: number, nullable: true, description: "Drawdown from 52w high as fraction" }
        trend3mPct:            { type: number, nullable: true, description: "3-month price trend as fraction" }
        ma50:                  { type: number, nullable: true }
        ma200:                 { type: number, nullable: true }
        panicDiscount:         { type: boolean, description: "True if price is in panic-buy territory" }
        deteriorationWarning:  { type: boolean, description: "True if price trend suggests structural deterioration" }
        seriesDays:            { type: integer, description: "Number of trading days in the underlying series" }
    VerdictBlock:
      type: object
      required: [verdettoClasse, positionSizePct, partialBasis, motivazioneAggregata, ruleCountGreen, ruleCountYellow, ruleCountRed, livelloRischio, newsSentimentDominante]
      description: Verdict cascade output (US-044).
      properties:
        verdettoClasse:          { $ref: '#/components/schemas/VerdictClass' }
        positionSizePct:         { type: number, description: "Suggested position size as percentage" }
        partialBasis:            { type: boolean, description: "True when LLM blocks were not invoked" }
        motivazioneAggregata:    { type: string, description: "Aggregated human-readable motivation" }
        ruleCountGreen:          { type: integer }
        ruleCountYellow:         { type: integer }
        ruleCountRed:            { type: integer }
        livelloRischio:          { $ref: '#/components/schemas/LivelloRischio' }
        newsSentimentDominante:  { $ref: '#/components/schemas/SentimentClass' }
    PositionSizeBlock:
      type: object
      required: [recommendedPct, rangeLow, rangeHigh, basisVerdict, marginOfSafetyPct, disclaimer]
      description: Kelly-inspired position sizing recommendation.
      properties:
        recommendedPct:    { type: number, description: "Central recommended allocation percentage" }
        rangeLow:          { type: number, description: "Conservative bound of allocation range" }
        rangeHigh:         { type: number, description: "Aggressive bound of allocation range" }
        basisVerdict:      { $ref: '#/components/schemas/VerdictClass' }
        marginOfSafetyPct: { type: number, description: "Margin of safety percentage used in sizing" }
        disclaimer:        { type: string }
    InversionItem:
      type: object
      required: [testo, chunkIndex]
      description: Single Munger inversion finding tied to a filing chunk.
      properties:
        testo:      { type: string, description: "Inversion finding text" }
        chunkIndex: { type: integer, description: "Index of the source filing chunk" }
    MungerReportBlock:
      type: object
      required: [livelloRischio, rischiPrincipali, puntiDiForza, segnaliRecenti10Q, filingComboHash, llmCallsCount]
      description: Munger inversion LLM report (US-041). Null when invoke_llm=false.
      properties:
        livelloRischio:    { $ref: '#/components/schemas/LivelloRischio' }
        rischiPrincipali:
          type: array
          items: { $ref: '#/components/schemas/InversionItem' }
        puntiDiForza:
          type: array
          items: { $ref: '#/components/schemas/InversionItem' }
        segnaliRecenti10Q:
          type: array
          items: { $ref: '#/components/schemas/InversionItem' }
        filingComboHash:   { type: string, description: "Hash of the filing combination used for cache keying" }
        llmCallsCount:     { type: integer }
    NewsSentimentBlock:
      type: object
      required: [total, panicCount, structuralCount, neutralCount, dominantClass]
      description: News sentiment classification summary (US-042). Null when invoke_llm=false.
      properties:
        total:           { type: integer }
        panicCount:      { type: integer }
        structuralCount: { type: integer }
        neutralCount:    { type: integer }
        dominantClass:   { $ref: '#/components/schemas/SentimentClass' }
    FilingRef:
      type: object
      required: [accessionNumber, formType, filingDate]
      description: Reference to a SEC filing used during deep analysis.
      properties:
        accessionNumber: { type: string }
        formType:        { type: string, description: "e.g. 10-K, 10-Q" }
        filingDate:      { type: string, description: "Filing date (YYYY-MM-DD)" }
    DeepAnalysisResponse:
      type: object
      required: [ticker, generatedAt, roe, priceAction, ruleEngineResults, verdict, filingsUsed, llmStatus, llmCalls, totalDurationMs]
      description: |
        Full response payload for GET /api/analysis/{ticker}/deep (US-045).
        Deterministic fields (roe, priceAction, ruleEngineResults, verdict,
        filingsUsed) are always populated. LLM-dependent fields (mungerReport,
        newsSentiment) are null when invoke_llm=false. positionSize is null
        when verdict is BOCCIATO_*.
      properties:
        ticker:      { type: string }
        generatedAt: { type: string, format: date-time }
        roe:
          $ref: '#/components/schemas/RoeBlock'
        priceAction:
          $ref: '#/components/schemas/PriceActionBlock'
        ruleEngineResults:
          type: array
          items: { $ref: '#/components/schemas/RuleSignal' }
          description: All 13 rule engine signals (7 Buffett + 6 Graham defensive)
        verdict:
          $ref: '#/components/schemas/VerdictBlock'
        positionSize:
          $ref: '#/components/schemas/PositionSizeBlock'
          nullable: true
          description: Position sizing recommendation, null for BOCCIATO verdicts
        filingsUsed:
          type: array
          items: { $ref: '#/components/schemas/FilingRef' }
        mungerReport:
          $ref: '#/components/schemas/MungerReportBlock'
          nullable: true
          description: Munger inversion LLM report, null when invoke_llm=false
        newsSentiment:
          $ref: '#/components/schemas/NewsSentimentBlock'
          nullable: true
          description: News sentiment classification, null when invoke_llm=false
        llmStatus:   { $ref: '#/components/schemas/LlmStatus' }
        llmCalls:    { type: integer, description: "Number of LLM API calls made" }
        totalDurationMs: { type: integer, format: int64, description: "Total pipeline duration in milliseconds" }

    # ---- EP-011 Deep Analysis — async run model (US-045, 2026-05-29) ----
    DeepAnalysisRunStatus:
      type: string
      description: Stato di un'esecuzione asincrona della deep analysis.
      enum: [RUNNING, SUCCESS, FAILED]
    LatestDeepAnalysisStatus:
      type: string
      description: |
        Stato dell'ultima esecuzione esposto da GET .../deep/latest. Aggiunge
        NONE rispetto a DeepAnalysisRunStatus per il caso "nessun run esistente".
      enum: [RUNNING, SUCCESS, FAILED, NONE]
    RunError:
      type: object
      required: [reason, message]
      description: Dettaglio errore di un run FAILED. reason = codice dominio (RFC 9457).
      properties:
        reason:  { type: string, description: "Codice motivo dominio (es. no_sec_filings, llm_unavailable, not_indexed). not_indexed = analisi LLM richiesta su un ticker mai indicizzato (lanciare prima POST .../deep/ingest)." }
        message: { type: string }
    DeepAnalysisRunKind:
      type: string
      description: |
        Tipo di operazione async sulla tabella deep_analysis_run (US-045,
        2026-05-30). INGEST = download filing SEC + embedding (no LLM); ANALYSIS =
        verdetto/analisi che riusa gli embedding persistiti.
      enum: [INGEST, ANALYSIS]
    DeepAnalysisRunStatusResponse:
      type: object
      required: [runId, ticker, status, invokeLlm, kind]
      description: |
        Risposta 202 di POST /api/analysis/{ticker}/deep/runs (kind=ANALYSIS) e di
        POST /api/analysis/{ticker}/deep/ingest (kind=INGEST, invokeLlm=false). In
        caso di dedupe runId è quello del run RUNNING già in corso per il ticker.
      properties:
        runId:     { type: string, format: uuid }
        ticker:    { type: string }
        status:    { $ref: '#/components/schemas/DeepAnalysisRunStatus' }
        invokeLlm: { type: boolean }
        kind:      { $ref: '#/components/schemas/DeepAnalysisRunKind' }
    IngestSummary:
      type: object
      required: [filingsTotal, chunksIndexed, chunksSkipped, indexedAt]
      description: |
        Esito di un INGEST riuscito. chunksSkipped > 0 quando alcuni filing erano
        già indicizzati (skip idempotente, nessun ri-embedding).
      properties:
        filingsTotal:  { type: integer, description: "Filing SEC considerati" }
        chunksIndexed: { type: integer, description: "Chunk nuovi calcolati ed embeddati" }
        chunksSkipped: { type: integer, description: "Chunk saltati perché già presenti in filing_chunks" }
        indexedAt:     { type: string, format: date-time, description: "Timestamp del completamento indicizzazione" }
    IngestStatusResponse:
      type: object
      required: [ticker, status]
      description: |
        Risposta 200 di GET /api/analysis/{ticker}/deep/ingest/latest. runId/
        requestedAt/completedAt sono null quando status=NONE. summary
        (IngestSummary) è valorizzato solo quando status=SUCCESS; error solo
        quando status=FAILED.
      properties:
        ticker:      { type: string }
        status:      { $ref: '#/components/schemas/LatestDeepAnalysisStatus' }
        runId:
          type: string
          format: uuid
          nullable: true
        requestedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        summary:
          allOf:
            - $ref: '#/components/schemas/IngestSummary'
          nullable: true
          description: Esito indicizzazione, presente solo quando status=SUCCESS
        error:
          allOf:
            - $ref: '#/components/schemas/RunError'
          nullable: true
          description: Dettaglio errore, presente solo quando status=FAILED
    LatestDeepAnalysisResponse:
      type: object
      required: [ticker, status, invokeLlm]
      description: |
        Risposta 200 di GET /api/analysis/{ticker}/deep/latest. runId/requestedAt/
        completedAt sono null quando status=NONE. result (DeepAnalysisResponse) è
        valorizzato solo quando status=SUCCESS; error solo quando status=FAILED.
      properties:
        ticker:      { type: string }
        status:      { $ref: '#/components/schemas/LatestDeepAnalysisStatus' }
        runId:
          type: string
          format: uuid
          nullable: true
        invokeLlm:   { type: boolean }
        requestedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        result:
          allOf:
            - $ref: '#/components/schemas/DeepAnalysisResponse'
          nullable: true
          description: Payload completo, presente solo quando status=SUCCESS
        error:
          allOf:
            - $ref: '#/components/schemas/RunError'
          nullable: true
          description: Dettaglio errore, presente solo quando status=FAILED

    # ---------------- EP-012 top value picks (US-050) ----------------
...